requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
mysql-connector-python
